# Conflicts:
#	backend/src/main/java/org/eduspace/backend/config/DataInitializer.java

+++ backend/src/main/java/org/eduspace/backend/dto/study_group/response/StudyGroupResponse.java
@@ -15,6 +15,6 @@ public class StudyGroupResponse {
     private Long studyGroupId;
     private List<GroupMemberInfo> members;
     private String status;
-    private Integer progress;
+    private Double progress;
 }
